move web3helpers.ts to test/ dir
